requirements.txt
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.9.0
numpy==2.2.5
openai==1.78.0
portalocker==2.10.1
protobuf==6.30.2
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
PyMySQL==1.1.1
qdrant-client==1.14.2
rest-framework-simplejwt==0.0.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
